extracts inline graphql queries from issueService
